+++ frontend/src/components/admin/order/AdminOrderManagement.tsx
@@ -18,6 +18,7 @@ import {
 import OrderDetailModal from "./OrderDetailModal";
 import OrderFilter from "./OrderFilter";
 import { Order } from "@/interface/Order";
+import { formatDateTime, isDateInRange } from "../../../utils/date/dateUtils";
 
 interface Props {
   orders: Order[];
@@ -28,20 +29,12 @@ const AdminOrderManagement: React.FC<Props> = ({ orders }) => {
   const [emailFilter, setEmailFilter] = useState("");
 
   const getDateFilteredOrders = () => {
-    const today = new Date();
-    const yesterday = new Date(today);
-    yesterday.setDate(yesterday.getDate() - 1);
-
-    today.setHours(14, 0, 0, 0);
-    yesterday.setHours(14, 0, 0, 0);
-
     return orders.filter((order) => {
-      const orderDate = new Date(order.ordered_at);
-
-      if (dateFilter === "today") {
-        return orderDate >= yesterday && orderDate <= today;
-      }
-      return true;
+      // 분리한 유틸리티 함수 사용
+      return isDateInRange(
+        order.ordered_at,
+        dateFilter === "today" ? "today" : "all"
+      );
     });
   };
 
@@ -94,7 +87,8 @@ const AdminOrderManagement: React.FC<Props> = ({ orders }) => {
                 <TableCell className="font-medium">{order.order_id}</TableCell>
                 <TableCell>{order.email}</TableCell>
                 <TableCell>
-                  {new Date(order.ordered_at).toLocaleString()}
+                  {/* 분리한 유틸리티 함수 사용 */}
+                  {formatDateTime(order.ordered_at)}
                 </TableCell>
                 <TableCell>
                   {order.order_products.reduce(
